Updated Accessory Serial Number Generation
I noticed that every now and then the persisted accessory could not be loaded.
This caused the switches to be regenerated, have the same SN assigned, and
in turn that caused my automations to fail.

Changed the serial number generation code to re-generate the same sn for
accessories with the same name.